Move toggleLogoDarkMode into domUpdates
