+++ src/domUpdates.js
@@ -1,4 +1,13 @@
 const domUpdates = {
+  toggleLogoDarkMode() {
+    const logo = document.querySelector('.logo-js')
+    logo.classList.toggle('dark-mode-on');
+    if (logo.classList.contains('dark-mode-on')) {
+      logo.src = './images/invertedCompass.png';
+    } else {
+      logo.src = './images/redBlueCompass.png';
+    }
+  }
   //innerHTML?
   //innerText?
   //insertAdjacentHTML?
