+++ src/scripts.js
@@ -7,8 +7,8 @@ import './images/arrow.png';
 
 
 import Traveler from './Traveler';
-import Trip from './Trip';
-import Destination from '.Destination';
+// import Trip from './Trip';
+// import Destination from '.Destination';
 import apiCalls from './apiCalls';
 import domUpdates from './domUpdates';
 
@@ -28,20 +28,11 @@ window.onload = Promise.all([apiCalls.getTravelersData(), apiCalls.getTripsData(
     data[2].destinations.forEach(destination => destinations.push(destination))
   })
 
+console.log(travelers);
 
 function toggleDarkMode() {
   document.querySelector('body').classList.toggle('dark-mode-on');
   darkLightModeButton.classList.toggle('dark-mode-on');
   nextDestinationButton.classList.toggle('dark-mode-on');
-  toggleLogoDarkMode();
-}
-
-function toggleLogoDarkMode() {
-  const logo = document.querySelector('.logo-js')
-  logo.classList.toggle('dark-mode-on');
-  if (logo.classList.contains('dark-mode-on')) {
-    logo.src = './images/invertedCompass.png';
-  } else {
-    logo.src = './images/redBlueCompass.png';
-  }
+  domUpdates.toggleLogoDarkMode();
 }
